centralize the screen view analytics for the Dashboard
