+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/home/HomeViewModel.kt
@@ -12,23 +12,16 @@ import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import org.cru.godtools.analytics.firebase.model.ACTION_IAM_HOME
-import org.cru.godtools.analytics.firebase.model.FirebaseIamActionEvent
-import org.cru.godtools.analytics.model.AnalyticsScreenEvent
-import org.cru.godtools.analytics.model.AnalyticsScreenEvent.Companion.SCREEN_HOME
 import org.cru.godtools.base.Settings
-import org.cru.godtools.base.ui.dashboard.Page
 import org.cru.godtools.model.Tool
 import org.cru.godtools.tutorial.PageSet
 import org.cru.godtools.ui.banner.BannerType
-import org.greenrobot.eventbus.EventBus
 import org.keynote.godtools.android.db.repository.LessonsRepository
 import org.keynote.godtools.android.db.repository.ToolsRepository
 
 @HiltViewModel
 @OptIn(ExperimentalCoroutinesApi::class)
 class HomeViewModel @Inject constructor(
-    private val eventBus: EventBus,
     lessonsRepository: LessonsRepository,
     settings: Settings,
     private val toolsRepository: ToolsRepository
@@ -45,16 +38,6 @@ class HomeViewModel @Inject constructor(
     val spotlightLessons = lessonsRepository.spotlightLessons
         .map { it.mapNotNull { it.code } }
         .stateIn(viewModelScope, SharingStarted.WhileSubscribed(), emptyList())
-
-    // region Analytics
-    fun trackPageInAnalytics(page: Page) = when (page) {
-        Page.HOME, Page.FAVORITE_TOOLS -> {
-            eventBus.post(AnalyticsScreenEvent(SCREEN_HOME))
-            eventBus.post(FirebaseIamActionEvent(ACTION_IAM_HOME))
-        }
-        else -> Unit
-    }
-    // endregion Analytics
 
     // region Favorites Tools
     val favoriteTools = toolsRepository.favoriteTools
