+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/lessons/LessonsLayout.kt
@@ -16,9 +16,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import org.ccci.gto.android.common.androidx.lifecycle.compose.OnResume
 import org.cru.godtools.R
-import org.cru.godtools.base.ui.dashboard.Page
 import org.cru.godtools.model.Tool
 import org.cru.godtools.model.Translation
 import org.cru.godtools.ui.tools.LessonToolCard
@@ -29,8 +27,6 @@ fun LessonsLayout(
     viewModel: LessonsViewModel = viewModel(),
     onOpenLesson: (Tool?, Translation?) -> Unit = { _, _ -> },
 ) {
-    OnResume { viewModel.trackPageInAnalytics(Page.LESSONS) }
-
     val lessons by viewModel.lessons.collectAsState()
     LazyColumn(contentPadding = PaddingValues(16.dp)) {
         item("header", "header") { LessonsHeader() }
