+++ app/src/main/kotlin/org/cru/godtools/ui/dashboard/tools/ToolsViewModel.kt
@@ -14,14 +14,9 @@ import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.flow.stateIn
 import org.ccci.gto.android.common.db.Expression.Companion.constants
 import org.ccci.gto.android.common.db.Query
-import org.cru.godtools.analytics.firebase.model.ACTION_IAM_ALL_TOOLS
-import org.cru.godtools.analytics.firebase.model.FirebaseIamActionEvent
-import org.cru.godtools.analytics.model.AnalyticsScreenEvent
 import org.cru.godtools.base.Settings
-import org.cru.godtools.base.ui.dashboard.Page
 import org.cru.godtools.model.Tool
 import org.cru.godtools.ui.banner.BannerType
-import org.greenrobot.eventbus.EventBus
 import org.keynote.godtools.android.db.Contract.ToolTable
 import org.keynote.godtools.android.db.GodToolsDao
 
@@ -45,7 +40,6 @@ internal val QUERY_TOOLS_SPOTLIGHT = QUERY_TOOLS_BASE.andWhere(ToolTable.FIELD_S
 @OptIn(ExperimentalCoroutinesApi::class)
 class ToolsViewModel @Inject constructor(
     dao: GodToolsDao,
-    private val eventBus: EventBus,
     settings: Settings,
     private val savedState: SavedStateHandle,
 ) : ViewModel() {
@@ -73,14 +67,4 @@ class ToolsViewModel @Inject constructor(
         tools.filter { category == null || it.category == category }
     }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000L), emptyList())
     // endregion Filters
-
-    // region Analytics
-    fun trackPageInAnalytics(page: Page) = when (page) {
-        Page.ALL_TOOLS -> {
-            eventBus.post(AnalyticsScreenEvent(AnalyticsScreenEvent.SCREEN_ALL_TOOLS))
-            eventBus.post(FirebaseIamActionEvent(ACTION_IAM_ALL_TOOLS))
-        }
-        else -> Unit
-    }
-    // endregion Analytics
 }
